third_party/blink/web_tests/platform/generic/media/media-fragments/TC0006-expected.txt


Title: TC0006
Fragment: 't=3,15'
Comment: The media is requested from a to e.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
